requirements.txt
numpy
pandas
nltk
flask
sqlite3
requests
beautifulsoup4
selenium
webdriver-manager
jinja2
python-dotenv==1.0.0
SQLAlchemy==2.0.20
pytest==7.4.0